Change acceptance criteria for cities + pigeons
Completed analysis that shows that you need 167 checklists to find all species in city that appear on at least 5% of all checklists.

So this changes criteria that a city musth have 167 checklists, and that we only include species as present if they appear on 5% of all checklists in a city.